requirements.txt
Flask
groq
python-dotenv
markdown2